Add better error handling when GH returns events
If an individual repo’s events call return value does not have a
message field, it is good, so add to allEvents.

If the message field exists and === ‘Not Found’ then warn and move onto
the next repo.  Otherwise, raise the error (which is most likely an API
request limit).

Currently, the widget completely fails to load if at least one repo is
mistyped.